[css,js] Popup look and thumbnail
In my.css: adds 13% transparency to popup-content-wrapper (for airy and
integrated feeling), adds 2px padding to popup elements (to center the
popup content and move the thumbnail further from [x] button), grows
thumbnail vertically by setting its max-height from 100px to 160px (for
a better view of the image), aligns location's text(li:nth-child(3)) on
left (to gain more space for the text).
In map.js: centers popup AND marker to the map + 23% of clientHeight (to
position it under the leaflet-control-locate icon on small
displays(<=800)), grows thumbnail horizontally by setting its max-width
from 100px to 160px (for a better view of the image), puts the
description in the same <ul> with title, artist, location and moves <br
clear="both"/> to the end of the function to allow wrapping along the
thumbnail (to integrate description's text in the text flow and the
whole layout).